WIP. Fix plans for tpch
