+++ modules/calcite/src/test/resources/tpch/q1.plan
@@ -1,5 +1,5 @@
-IgniteReduceSortAggregate(rowType=[RecordType(VARCHAR L_RETURNFLAG, VARCHAR L_LINESTATUS, DECIMAL(32767, 0) SUM_QTY, DECIMAL(32767, 0) SUM_BASE_PRICE, DECIMAL(32767, 0) SUM_DISC_PRICE, DECIMAL(32767, 0) SUM_CHARGE, DECIMAL(15, 2) AVG_QTY, DECIMAL(15, 2) AVG_PRICE, DECIMAL(15, 2) AVG_DISC, BIGINT COUNT_ORDER)], group=[{0, 1}], SUM_QTY=[SUM($2)], SUM_BASE_PRICE=[SUM($3)], SUM_DISC_PRICE=[SUM($4)], SUM_CHARGE=[SUM($5)], AVG_QTY=[AVG($2)], AVG_PRICE=[AVG($3)], AVG_DISC=[AVG($6)], COUNT_ORDER=[COUNT()], collation=[[0 ASC-nulls-first, 1 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=8.0, memory=77.0, io=1.0, network=13.0], id = 117
-  IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=7.0, memory=77.0, io=1.0, network=13.0], id = 116
-    IgniteMapSortAggregate(group=[{0, 1}], SUM_QTY=[SUM($2)], SUM_BASE_PRICE=[SUM($3)], SUM_DISC_PRICE=[SUM($4)], SUM_CHARGE=[SUM($5)], AVG_QTY=[AVG($2)], AVG_PRICE=[AVG($3)], AVG_DISC=[AVG($6)], COUNT_ORDER=[COUNT()], collation=[[0 ASC-nulls-first, 1 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=77.0, io=1.0, network=1.0], id = 115
-      IgniteSort(sort0=[$0], sort1=[$1], dir0=[ASC-nulls-first], dir1=[ASC-nulls-first]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=29.0, io=1.0, network=1.0], id = 114
-        IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_SD], filters=[<=($t6, 1998-09-02)], rowType=[RecordType(VARCHAR L_RETURNFLAG, VARCHAR L_LINESTATUS, DECIMAL(15, 2) L_QUANTITY, DECIMAL(15, 2) L_EXTENDEDPRICE, DECIMAL(31, 4) $f4, DECIMAL(47, 6) $f5, DECIMAL(15, 2) L_DISCOUNT)], projects=[[$t4, $t5, $t0, $t1, *($t1, -(1, $t2)), *(*($t1, -(1, $t2)), +(1, $t3)), $t2]], requiredColumns=[{6, 7, 8, 9, 10, 11, 12}], searchBounds=[[null, null, null, null, null, null, null, null, null, null, null, null, RangeBounds [lowerBound=$NULL_BOUND(), upperBound=1998-09-02, lowerInclude=false, upperInclude=true], null, null, null, null, null]], inlineScan=[false], collation=[[12 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 75
+IgniteReduceSortAggregate(rowType=[RecordType(VARCHAR L_RETURNFLAG, VARCHAR L_LINESTATUS, DECIMAL(32767, 0) SUM_QTY, DECIMAL(32767, 0) SUM_BASE_PRICE, DECIMAL(32767, 0) SUM_DISC_PRICE, DECIMAL(32767, 0) SUM_CHARGE, DECIMAL(15, 2) AVG_QTY, DECIMAL(15, 2) AVG_PRICE, DECIMAL(15, 2) AVG_DISC, BIGINT COUNT_ORDER)], group=[{0, 1}], SUM_QTY=[SUM($2)], SUM_BASE_PRICE=[SUM($3)], SUM_DISC_PRICE=[SUM($4)], SUM_CHARGE=[SUM($5)], AVG_QTY=[AVG($2)], AVG_PRICE=[AVG($3)], AVG_DISC=[AVG($6)], COUNT_ORDER=[COUNT()], collation=[[0 ASC-nulls-first, 1 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=8.0, memory=77.0, io=1.0, network=13.0], id = {id}
+  IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=7.0, memory=77.0, io=1.0, network=13.0], id = {id}
+    IgniteMapSortAggregate(group=[{0, 1}], SUM_QTY=[SUM($2)], SUM_BASE_PRICE=[SUM($3)], SUM_DISC_PRICE=[SUM($4)], SUM_CHARGE=[SUM($5)], AVG_QTY=[AVG($2)], AVG_PRICE=[AVG($3)], AVG_DISC=[AVG($6)], COUNT_ORDER=[COUNT()], collation=[[0 ASC-nulls-first, 1 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=77.0, io=1.0, network=1.0], id = {id}
+      IgniteSort(sort0=[$0], sort1=[$1], dir0=[ASC-nulls-first], dir1=[ASC-nulls-first]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=29.0, io=1.0, network=1.0], id = {id}
+        IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_SD], filters=[<=($t6, 1998-09-02)], rowType=[RecordType(VARCHAR L_RETURNFLAG, VARCHAR L_LINESTATUS, DECIMAL(15, 2) L_QUANTITY, DECIMAL(15, 2) L_EXTENDEDPRICE, DECIMAL(31, 4) $f4, DECIMAL(47, 6) $f5, DECIMAL(15, 2) L_DISCOUNT)], projects=[[$t4, $t5, $t0, $t1, *($t1, -(1, $t2)), *(*($t1, -(1, $t2)), +(1, $t3)), $t2]], requiredColumns=[{6, 7, 8, 9, 10, 11, 12}], searchBounds=[[null, null, null, null, null, null, null, null, null, null, null, null, RangeBounds [lowerBound=$NULL_BOUND(), upperBound=1998-09-02, lowerInclude=false, upperInclude=true], null, null, null, null, null]], inlineScan=[false], collation=[[12 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
