+++ modules/calcite/src/test/resources/tpch/q12.plan
@@ -1,9 +1,9 @@
-IgniteReduceSortAggregate(rowType=[RecordType(VARCHAR L_SHIPMODE, BIGINT HIGH_LINE_COUNT, BIGINT LOW_LINE_COUNT)], group=[{0}], HIGH_LINE_COUNT=[SUM($1)], LOW_LINE_COUNT=[SUM($2)], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.0, cpu=15.0, memory=24.0, io=2.0, network=18.0], id = 54777
-  IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=8.0, cpu=14.0, memory=24.0, io=2.0, network=18.0], id = 54776
-    IgniteMapSortAggregate(group=[{0}], HIGH_LINE_COUNT=[SUM($1)], LOW_LINE_COUNT=[SUM($2)], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=7.0, cpu=13.0, memory=24.0, io=2.0, network=10.0], id = 54775
-      IgniteProject(L_SHIPMODE=[$1], $f1=[CASE(SEARCH($3, Sarg[_UTF-8'1-URGENT':VARCHAR CHARACTER SET "UTF-8", _UTF-8'2-HIGH':VARCHAR CHARACTER SET "UTF-8"]:VARCHAR CHARACTER SET "UTF-8"), 1, 0)], $f2=[CASE(SEARCH($3, Sarg[(-∞.._UTF-8'1-URGENT':VARCHAR CHARACTER SET "UTF-8"), (_UTF-8'1-URGENT':VARCHAR CHARACTER SET "UTF-8".._UTF-8'2-HIGH':VARCHAR CHARACTER SET "UTF-8"), (_UTF-8'2-HIGH':VARCHAR CHARACTER SET "UTF-8"..+∞)]:VARCHAR CHARACTER SET "UTF-8"), 1, 0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.0, cpu=12.0, memory=10.0, io=2.0, network=10.0], id = 54774
-        IgniteCorrelatedNestedLoopJoin(condition=[=($2, $0)], joinType=[inner], variablesSet=[[$cor0]], variablesSet=[[0]], correlationVariables=[[$cor0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=11.0, memory=10.0, io=2.0, network=10.0], id = 54773
-          IgniteExchange(distribution=[affinity[identity=AffinityIdentity [affFuncCls=class org.apache.ignite.cache.affinity.rendezvous.RendezvousAffinityFunction, backups=0, partsCnt=1024, filterCls=class org.apache.ignite.configuration.CacheConfiguration$IgniteAllNodesPredicate, hash=-39424137], cacheId=-1464737622][0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=9.0, io=1.0, network=9.0], id = 54772
-            IgniteSort(sort0=[$1], dir0=[ASC-nulls-first]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=9.0, io=1.0, network=1.0], id = 54771
-              IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_RD], filters=[AND(OR(=($t4, _UTF-8'MAIL'), =($t4, _UTF-8'SHIP')), <($t2, $t3), <($t1, $t2), AND(>=($t3, 1994-01-01), <($t3, 1995-01-01)))], rowType=[RecordType(INTEGER L_ORDERKEY, VARCHAR L_SHIPMODE)], projects=[[$t0, $t4]], requiredColumns=[{2, 12, 13, 14, 16}], searchBounds=[[null, null, null, null, null, null, null, null, null, null, null, null, null, null, RangeBounds [lowerBound=1994-01-01, upperBound=1995-01-01, lowerInclude=true, upperInclude=false], null, null, null]], inlineScan=[false], collation=[[14 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 47453
-          IgniteIndexScan(table=[[PUBLIC, ORDERS]], index=[{INDEX_CASE:_key_PK,_key_PK_proxy}], filters=[=($t0, $cor0.L_ORDERKEY)], requiredColumns=[{2, 7}], searchBounds=[[null, null, ExactBounds [bound=$cor0.L_ORDERKEY], null, null, null, null, null, null, null, null]], inlineScan=[false], collation=[[2 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 49224
+IgniteReduceSortAggregate(rowType=[RecordType(VARCHAR L_SHIPMODE, BIGINT HIGH_LINE_COUNT, BIGINT LOW_LINE_COUNT)], group=[{0}], HIGH_LINE_COUNT=[SUM($1)], LOW_LINE_COUNT=[SUM($2)], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.0, cpu=15.0, memory=24.0, io=2.0, network=18.0], id = {id}
+  IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=8.0, cpu=14.0, memory=24.0, io=2.0, network=18.0], id = {id}
+    IgniteMapSortAggregate(group=[{0}], HIGH_LINE_COUNT=[SUM($1)], LOW_LINE_COUNT=[SUM($2)], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=7.0, cpu=13.0, memory=24.0, io=2.0, network=10.0], id = {id}
+      IgniteProject(L_SHIPMODE=[$1], $f1=[CASE(SEARCH($3, Sarg[_UTF-8'1-URGENT':VARCHAR CHARACTER SET "UTF-8", _UTF-8'2-HIGH':VARCHAR CHARACTER SET "UTF-8"]:VARCHAR CHARACTER SET "UTF-8"), 1, 0)], $f2=[CASE(SEARCH($3, Sarg[(-∞.._UTF-8'1-URGENT':VARCHAR CHARACTER SET "UTF-8"), (_UTF-8'1-URGENT':VARCHAR CHARACTER SET "UTF-8".._UTF-8'2-HIGH':VARCHAR CHARACTER SET "UTF-8"), (_UTF-8'2-HIGH':VARCHAR CHARACTER SET "UTF-8"..+∞)]:VARCHAR CHARACTER SET "UTF-8"), 1, 0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.0, cpu=12.0, memory=10.0, io=2.0, network=10.0], id = {id}
+        IgniteCorrelatedNestedLoopJoin(condition=[=($2, $0)], joinType=[inner], variablesSet=[[$cor0]], variablesSet=[[0]], correlationVariables=[[$cor0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=11.0, memory=10.0, io=2.0, network=10.0], id = {id}
+          IgniteExchange(distribution=[affinity[identity=AffinityIdentity [affFuncCls=class org.apache.ignite.cache.affinity.rendezvous.RendezvousAffinityFunction, backups=0, partsCnt=1024, filterCls=class org.apache.ignite.configuration.CacheConfiguration$IgniteAllNodesPredicate, hash={hash}, cacheId=-1464737622][0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=9.0, io=1.0, network=9.0], id = {id}
+            IgniteSort(sort0=[$1], dir0=[ASC-nulls-first]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=9.0, io=1.0, network=1.0], id = {id}
+              IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_RD], filters=[AND(OR(=($t4, _UTF-8'MAIL'), =($t4, _UTF-8'SHIP')), <($t2, $t3), <($t1, $t2), AND(>=($t3, 1994-01-01), <($t3, 1995-01-01)))], rowType=[RecordType(INTEGER L_ORDERKEY, VARCHAR L_SHIPMODE)], projects=[[$t0, $t4]], requiredColumns=[{2, 12, 13, 14, 16}], searchBounds=[[null, null, null, null, null, null, null, null, null, null, null, null, null, null, RangeBounds [lowerBound=1994-01-01, upperBound=1995-01-01, lowerInclude=true, upperInclude=false], null, null, null]], inlineScan=[false], collation=[[14 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
+          IgniteIndexScan(table=[[PUBLIC, ORDERS]], index=[{INDEX_CASE:_key_PK,_key_PK_proxy}], filters=[=($t0, $cor0.L_ORDERKEY)], requiredColumns=[{2, 7}], searchBounds=[[null, null, ExactBounds [bound=$cor0.L_ORDERKEY], null, null, null, null, null, null, null, null]], inlineScan=[false], collation=[[2 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
