+++ modules/calcite/src/test/resources/tpch/q15.plan
@@ -1,14 +1,14 @@
-IgniteProject(S_SUPPKEY=[$0], S_NAME=[$1], S_ADDRESS=[$2], S_PHONE=[$3], TOTAL_REVENUE=[$6]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=16.0, cpu=28.0, memory=25.5, io=3.0, network=35.0], id = 332086
-  IgniteMergeJoin(condition=[=($0, $5)], joinType=[inner], variablesSet=[[]], leftCollation=[[0 ASC-nulls-first]], rightCollation=[[1 ASC-nulls-first]], allowNulls=[{}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=15.0, cpu=27.0, memory=25.5, io=3.0, network=35.0], id = 332085
-    IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=17.0], id = 332076
-      IgniteIndexScan(table=[[PUBLIC, SUPPLIER]], index=[{INDEX_CASE:_key_PK,_key_PK_proxy}], requiredColumns=[{2, 3, 4, 6}], inlineScan=[false], collation=[[2 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 64355
-    IgniteProject(EXPR$0=[$2], L_SUPPKEY=[$0], EXPR$1=[$1]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=11.0, cpu=17.0, memory=24.5, io=2.0, network=18.0], id = 332084
-      IgniteNestedLoopJoin(condition=[=($1, $2)], joinType=[inner], variablesSet=[[]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.0, cpu=16.0, memory=24.5, io=2.0, network=18.0], id = 332083
-        IgniteColocatedSortAggregate(group=[{0}], EXPR$1=[SUM($1)], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=7.0, memory=10.0, io=1.0, network=9.0], id = 332078
-          IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=1.0, io=1.0, network=9.0], id = 332077
-            IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_SK_PK], filters=[AND(>=($t3, 1996-01-01), <($t3, 1996-04-01))], rowType=[RecordType(INTEGER L_SUPPKEY, DECIMAL(31, 4) $f1)], projects=[[$t0, *($t1, -(1, $t2))]], requiredColumns=[{4, 7, 8, 12}], inlineScan=[false], collation=[[4 ASC-nulls-first, 3 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=1.0, io=1.0, network=1.0], id = 63108
-        IgniteColocatedHashAggregate(group=[{}], EXPR$0=[MAX($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=5.0, memory=10.5, io=1.0, network=9.0], id = 332082
-          IgniteProject(EXPR$1=[$1]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=4.0, memory=5.5, io=1.0, network=9.0], id = 332081
-            IgniteColocatedHashAggregate(group=[{0}], EXPR$1=[SUM($1)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=3.0, memory=5.5, io=1.0, network=9.0], id = 332080
-              IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=9.0], id = 332079
-                IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_SD], filters=[AND(>=($t3, 1996-01-01), <($t3, 1996-04-01))], rowType=[RecordType(INTEGER L_SUPPKEY, DECIMAL(31, 4) $f1)], projects=[[$t0, *($t1, -(1, $t2))]], requiredColumns=[{4, 7, 8, 12}], searchBounds=[[null, null, null, null, null, null, null, null, null, null, null, null, RangeBounds [lowerBound=1996-01-01, upperBound=1996-04-01, lowerInclude=true, upperInclude=false], null, null, null, null, null]], inlineScan=[false], collation=[[12 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 63103
+IgniteProject(S_SUPPKEY=[$0], S_NAME=[$1], S_ADDRESS=[$2], S_PHONE=[$3], TOTAL_REVENUE=[$6]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=16.0, cpu=28.0, memory=25.5, io=3.0, network=35.0], id = {id}
+  IgniteMergeJoin(condition=[=($0, $5)], joinType=[inner], variablesSet=[[]], leftCollation=[[0 ASC-nulls-first]], rightCollation=[[1 ASC-nulls-first]], allowNulls=[{}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=15.0, cpu=27.0, memory=25.5, io=3.0, network=35.0], id = {id}
+    IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=17.0], id = {id}
+      IgniteIndexScan(table=[[PUBLIC, SUPPLIER]], index=[{INDEX_CASE:_key_PK,_key_PK_proxy}], requiredColumns=[{2, 3, 4, 6}], inlineScan=[false], collation=[[2 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
+    IgniteProject(EXPR$0=[$2], L_SUPPKEY=[$0], EXPR$1=[$1]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=11.0, cpu=17.0, memory=24.5, io=2.0, network=18.0], id = {id}
+      IgniteNestedLoopJoin(condition=[=($1, $2)], joinType=[inner], variablesSet=[[]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.0, cpu=16.0, memory=24.5, io=2.0, network=18.0], id = {id}
+        IgniteColocatedSortAggregate(group=[{0}], EXPR$1=[SUM($1)], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=7.0, memory=10.0, io=1.0, network=9.0], id = {id}
+          IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=1.0, io=1.0, network=9.0], id = {id}
+            IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_SK_PK], filters=[AND(>=($t3, 1996-01-01), <($t3, 1996-04-01))], rowType=[RecordType(INTEGER L_SUPPKEY, DECIMAL(31, 4) $f1)], projects=[[$t0, *($t1, -(1, $t2))]], requiredColumns=[{4, 7, 8, 12}], inlineScan=[false], collation=[[4 ASC-nulls-first, 3 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=1.0, io=1.0, network=1.0], id = {id}
+        IgniteColocatedHashAggregate(group=[{}], EXPR$0=[MAX($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=5.0, memory=10.5, io=1.0, network=9.0], id = {id}
+          IgniteProject(EXPR$1=[$1]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=4.0, memory=5.5, io=1.0, network=9.0], id = {id}
+            IgniteColocatedHashAggregate(group=[{0}], EXPR$1=[SUM($1)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=3.0, memory=5.5, io=1.0, network=9.0], id = {id}
+              IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=9.0], id = {id}
+                IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_SD], filters=[AND(>=($t3, 1996-01-01), <($t3, 1996-04-01))], rowType=[RecordType(INTEGER L_SUPPKEY, DECIMAL(31, 4) $f1)], projects=[[$t0, *($t1, -(1, $t2))]], requiredColumns=[{4, 7, 8, 12}], searchBounds=[[null, null, null, null, null, null, null, null, null, null, null, null, RangeBounds [lowerBound=1996-01-01, upperBound=1996-04-01, lowerInclude=true, upperInclude=false], null, null, null, null, null]], inlineScan=[false], collation=[[12 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
