+++ modules/calcite/src/test/resources/tpch/q17.plan
@@ -1,16 +1,16 @@
-IgniteProject(AVG_YEARLY=[/($0, 7.0:DECIMAL(2, 1))]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.9, cpu=19.75, memory=13.1, io=2.15, network=19.35], id = 384863
-  IgniteColocatedHashAggregate(group=[{}], agg#0=[SUM($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.9, cpu=18.75, memory=13.1, io=2.15, network=19.35], id = 384862
-    IgniteProject(L_EXTENDEDPRICE=[$2]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=8.9, cpu=17.75, memory=8.1, io=2.15, network=19.35], id = 384861
-      IgniteFilter(condition=[<(CAST($1):DECIMAL(17, 3) NOT NULL, $4)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=7.9, cpu=16.75, memory=8.1, io=2.15, network=19.35], id = 384860
-        IgniteCorrelatedNestedLoopJoin(condition=[true], joinType=[left], variablesSet=[[$cor0]], variablesSet=[[0]], correlationVariables=[[$cor0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.9, cpu=12.75, memory=8.1, io=2.15, network=19.35], id = 384859
-          IgniteNestedLoopJoin(condition=[=($3, $0)], joinType=[inner], variablesSet=[[]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=8.0, memory=6.0, io=2.0, network=18.0], id = 384853
-            IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=13.0], id = 384851
-              IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_PK_SK], requiredColumns=[{3, 6, 7}], inlineScan=[false], collation=[[3 ASC-nulls-first, 4 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 381798
-            IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=5.0], id = 384852
-              IgniteTableScan(table=[[PUBLIC, PART]], filters=[AND(=($t1, _UTF-8'Brand#23'), =($t2, _UTF-8'MED BOX'))], rowType=[RecordType(INTEGER P_PARTKEY)], projects=[[$t0]], requiredColumns=[{2, 5, 8}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 381776
-          IgniteProject(EXPR$0=[*(0.2:DECIMAL(2, 1), $0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.0, cpu=5.0, memory=14.0, io=1.0, network=9.0], id = 384858
-            IgniteColocatedHashAggregate(group=[{}], agg#0=[AVG($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=4.0, memory=14.0, io=1.0, network=9.0], id = 384857
-              IgniteProject(L_QUANTITY=[$1]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=3.0, memory=9.0, io=1.0, network=9.0], id = 384856
-                IgniteSortedIndexSpool(readType=[LAZY], writeType=[EAGER], condition=[=($0, $cor0.P_PARTKEY)], collation=[[0 ASC-nulls-first]], searchBounds=[[ExactBounds [bound=$cor0.P_PARTKEY], null]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=2.0, memory=9.0, io=1.0, network=9.0], id = 384855
-                  IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=9.0], id = 384854
-                    IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_PK_SK], requiredColumns=[{3, 6}], inlineScan=[false], collation=[[3 ASC-nulls-first, 4 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 383774
+IgniteProject(AVG_YEARLY=[/($0, 7.0:DECIMAL(2, 1))]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.9, cpu=19.75, memory=13.1, io=2.15, network=19.35], id = {id}
+  IgniteColocatedHashAggregate(group=[{}], agg#0=[SUM($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.9, cpu=18.75, memory=13.1, io=2.15, network=19.35], id = {id}
+    IgniteProject(L_EXTENDEDPRICE=[$2]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=8.9, cpu=17.75, memory=8.1, io=2.15, network=19.35], id = {id}
+      IgniteFilter(condition=[<(CAST($1):DECIMAL(17, 3) NOT NULL, $4)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=7.9, cpu=16.75, memory=8.1, io=2.15, network=19.35], id = {id}
+        IgniteCorrelatedNestedLoopJoin(condition=[true], joinType=[left], variablesSet=[[$cor0]], variablesSet=[[0]], correlationVariables=[[$cor0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.9, cpu=12.75, memory=8.1, io=2.15, network=19.35], id = {id}
+          IgniteNestedLoopJoin(condition=[=($3, $0)], joinType=[inner], variablesSet=[[]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=8.0, memory=6.0, io=2.0, network=18.0], id = {id}
+            IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=13.0], id = {id}
+              IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_PK_SK], requiredColumns=[{3, 6, 7}], inlineScan=[false], collation=[[3 ASC-nulls-first, 4 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
+            IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=5.0], id = {id}
+              IgniteTableScan(table=[[PUBLIC, PART]], filters=[AND(=($t1, _UTF-8'Brand#23'), =($t2, _UTF-8'MED BOX'))], rowType=[RecordType(INTEGER P_PARTKEY)], projects=[[$t0]], requiredColumns=[{2, 5, 8}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
+          IgniteProject(EXPR$0=[*(0.2:DECIMAL(2, 1), $0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.0, cpu=5.0, memory=14.0, io=1.0, network=9.0], id = {id}
+            IgniteColocatedHashAggregate(group=[{}], agg#0=[AVG($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=4.0, memory=14.0, io=1.0, network=9.0], id = {id}
+              IgniteProject(L_QUANTITY=[$1]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=3.0, memory=9.0, io=1.0, network=9.0], id = {id}
+                IgniteSortedIndexSpool(readType=[LAZY], writeType=[EAGER], condition=[=($0, $cor0.P_PARTKEY)], collation=[[0 ASC-nulls-first]], searchBounds=[[ExactBounds [bound=$cor0.P_PARTKEY], null]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=2.0, memory=9.0, io=1.0, network=9.0], id = {id}
+                  IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=9.0], id = {id}
+                    IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_PK_SK], requiredColumns=[{3, 6}], inlineScan=[false], collation=[[3 ASC-nulls-first, 4 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
