+++ modules/calcite/src/test/resources/tpch/q22.plan
@@ -1,20 +1,20 @@
-IgniteColocatedSortAggregate(group=[{0}], NUMCUST=[COUNT()], TOTACCTBAL=[SUM($1)], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=13.5, cpu=25.0, memory=37.5, io=2.5, network=20.5], id = 553106
-  IgniteSort(sort0=[$0], dir0=[ASC-nulls-first]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=12.5, cpu=24.0, memory=23.5, io=2.5, network=20.5], id = 553105
-    IgniteProject(CNTRYCODE=[SUBSTR($1, 1, 2)], C_ACCTBAL=[$2]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=11.5, cpu=20.0, memory=15.5, io=2.5, network=20.5], id = 553104
-      IgniteFilter(condition=[IS NULL($4)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.5, cpu=19.0, memory=15.5, io=2.5, network=20.5], id = 553103
-        IgniteCorrelatedNestedLoopJoin(condition=[true], joinType=[left], variablesSet=[[$cor0]], variablesSet=[[0]], correlationVariables=[[$cor0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.5, cpu=15.0, memory=15.5, io=2.5, network=20.5], id = 553102
-          IgniteNestedLoopJoin(condition=[>($2, $3)], joinType=[inner], variablesSet=[[]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.0, cpu=9.0, memory=11.0, io=2.0, network=18.0], id = 553097
-            IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=13.0], id = 553094
-              IgniteTableScan(table=[[PUBLIC, CUSTOMER]], filters=[OR(=(SUBSTR($t1, 1, 2), _UTF-8'13'), =(SUBSTR($t1, 1, 2), _UTF-8'31'), =(SUBSTR($t1, 1, 2), _UTF-8'23'), =(SUBSTR($t1, 1, 2), _UTF-8'29'), =(SUBSTR($t1, 1, 2), _UTF-8'30'), =(SUBSTR($t1, 1, 2), _UTF-8'18'), =(SUBSTR($t1, 1, 2), _UTF-8'17'))], requiredColumns=[{2, 6, 7}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 551950
-            IgniteColocatedHashAggregate(group=[{}], EXPR$0=[AVG($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=3.0, memory=6.0, io=1.0, network=5.0], id = 553096
-              IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=5.0], id = 553095
-                IgniteTableScan(table=[[PUBLIC, CUSTOMER]], filters=[AND(>($t1, 0.00), OR(=(SUBSTR($t0, 1, 2), _UTF-8'13'), =(SUBSTR($t0, 1, 2), _UTF-8'31'), =(SUBSTR($t0, 1, 2), _UTF-8'23'), =(SUBSTR($t0, 1, 2), _UTF-8'29'), =(SUBSTR($t0, 1, 2), _UTF-8'30'), =(SUBSTR($t0, 1, 2), _UTF-8'18'), =(SUBSTR($t0, 1, 2), _UTF-8'17')))], rowType=[RecordType(DECIMAL(15, 2) C_ACCTBAL)], projects=[[$t1]], requiredColumns=[{6, 7}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 551886
-          IgniteColocatedHashAggregate(group=[{0}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=4.0, memory=9.0, io=1.0, network=5.0], id = 553101
-            IgniteProject(i=[true]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=3.0, memory=5.0, io=1.0, network=5.0], id = 553100
-              IgniteSortedIndexSpool(readType=[LAZY], writeType=[EAGER], condition=[=($0, $cor0.C_CUSTKEY)], collation=[[0 ASC-nulls-first]], searchBounds=[[ExactBounds [bound=$cor0.C_CUSTKEY]]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=2.0, memory=5.0, io=1.0, network=5.0], id = 553099
-                IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=5.0], id = 553098
+IgniteColocatedSortAggregate(group=[{0}], NUMCUST=[COUNT()], TOTACCTBAL=[SUM($1)], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=13.5, cpu=25.0, memory=37.5, io=2.5, network=20.5], id = {id}
+  IgniteSort(sort0=[$0], dir0=[ASC-nulls-first]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=12.5, cpu=24.0, memory=23.5, io=2.5, network=20.5], id = {id}
+    IgniteProject(CNTRYCODE=[SUBSTR($1, 1, 2)], C_ACCTBAL=[$2]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=11.5, cpu=20.0, memory=15.5, io=2.5, network=20.5], id = {id}
+      IgniteFilter(condition=[IS NULL($4)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.5, cpu=19.0, memory=15.5, io=2.5, network=20.5], id = {id}
+        IgniteCorrelatedNestedLoopJoin(condition=[true], joinType=[left], variablesSet=[[$cor0]], variablesSet=[[0]], correlationVariables=[[$cor0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.5, cpu=15.0, memory=15.5, io=2.5, network=20.5], id = {id}
+          IgniteNestedLoopJoin(condition=[>($2, $3)], joinType=[inner], variablesSet=[[]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.0, cpu=9.0, memory=11.0, io=2.0, network=18.0], id = {id}
+            IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=13.0], id = {id}
+              IgniteTableScan(table=[[PUBLIC, CUSTOMER]], filters=[OR(=(SUBSTR($t1, 1, 2), _UTF-8'13'), =(SUBSTR($t1, 1, 2), _UTF-8'31'), =(SUBSTR($t1, 1, 2), _UTF-8'23'), =(SUBSTR($t1, 1, 2), _UTF-8'29'), =(SUBSTR($t1, 1, 2), _UTF-8'30'), =(SUBSTR($t1, 1, 2), _UTF-8'18'), =(SUBSTR($t1, 1, 2), _UTF-8'17'))], requiredColumns=[{2, 6, 7}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
+            IgniteColocatedHashAggregate(group=[{}], EXPR$0=[AVG($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=3.0, memory=6.0, io=1.0, network=5.0], id = {id}
+              IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=5.0], id = {id}
+                IgniteTableScan(table=[[PUBLIC, CUSTOMER]], filters=[AND(>($t1, 0.00), OR(=(SUBSTR($t0, 1, 2), _UTF-8'13'), =(SUBSTR($t0, 1, 2), _UTF-8'31'), =(SUBSTR($t0, 1, 2), _UTF-8'23'), =(SUBSTR($t0, 1, 2), _UTF-8'29'), =(SUBSTR($t0, 1, 2), _UTF-8'30'), =(SUBSTR($t0, 1, 2), _UTF-8'18'), =(SUBSTR($t0, 1, 2), _UTF-8'17')))], rowType=[RecordType(DECIMAL(15, 2) C_ACCTBAL)], projects=[[$t1]], requiredColumns=[{6, 7}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
+          IgniteColocatedHashAggregate(group=[{0}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=4.0, memory=9.0, io=1.0, network=5.0], id = {id}
+            IgniteProject(i=[true]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=3.0, memory=5.0, io=1.0, network=5.0], id = {id}
+              IgniteSortedIndexSpool(readType=[LAZY], writeType=[EAGER], condition=[=($0, $cor0.C_CUSTKEY)], collation=[[0 ASC-nulls-first]], searchBounds=[[ExactBounds [bound=$cor0.C_CUSTKEY]]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=2.0, memory=5.0, io=1.0, network=5.0], id = {id}
+                IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=5.0], id = {id}
 {ONEOF}
-                  IgniteIndexScan(table=[[PUBLIC, ORDERS]], index=[O_CK], requiredColumns=[{3}], inlineScan=[true], collation=[[3 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 552155
+                  IgniteIndexScan(table=[[PUBLIC, ORDERS]], index=[O_CK], requiredColumns=[{3}], inlineScan=[true], collation=[[3 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
 =====
                   IgniteIndexScan(table=[[PUBLIC, ORDERS]], index=[O_CK_proxy], requiredColumns=[{3}], inlineScan=[true], collation=[[3 ASC-nulls-first, 2 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
 {ONEOF}
