+++ modules/calcite/src/test/resources/tpch/q4.plan
@@ -1,11 +1,11 @@
-IgniteColocatedSortAggregate(group=[{0}], ORDER_COUNT=[COUNT()], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=12.0, cpu=20.0, memory=23.0, io=2.0, network=14.0], id = 889723
-  IgniteSort(sort0=[$0], dir0=[ASC-nulls-first]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=11.0, cpu=19.0, memory=14.0, io=2.0, network=14.0], id = 889722
-    IgniteProject(O_ORDERPRIORITY=[$1]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.0, cpu=15.0, memory=10.0, io=2.0, network=14.0], id = 889721
-      IgniteCorrelatedNestedLoopJoin(condition=[true], joinType=[inner], variablesSet=[[$cor0]], variablesSet=[[0]], correlationVariables=[[$cor0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.0, cpu=14.0, memory=10.0, io=2.0, network=14.0], id = 889720
-        IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=9.0], id = 889715
-          IgniteIndexScan(table=[[PUBLIC, ORDERS]], index=[O_OD], filters=[AND(>=($t1, 1993-07-01), <($t1, 1993-10-01))], rowType=[RecordType(INTEGER O_ORDERKEY, VARCHAR O_ORDERPRIORITY)], projects=[[$t0, $t2]], requiredColumns=[{2, 6, 7}], searchBounds=[[null, null, null, null, null, null, RangeBounds [lowerBound=1993-07-01, upperBound=1993-10-01, lowerInclude=true, upperInclude=false], null, null, null, null]], inlineScan=[false], collation=[[6 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 889033
-        IgniteColocatedHashAggregate(group=[{0}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.0, cpu=8.0, memory=9.0, io=1.0, network=5.0], id = 889719
-          IgniteProject(i=[true]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=7.0, memory=5.0, io=1.0, network=5.0], id = 889718
-            IgniteSortedIndexSpool(readType=[LAZY], writeType=[EAGER], condition=[=($0, $cor0.O_ORDERKEY)], collation=[[0 ASC-nulls-first]], searchBounds=[[ExactBounds [bound=$cor0.O_ORDERKEY]]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=6.0, memory=5.0, io=1.0, network=5.0], id = 889717
-              IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=1.0, io=1.0, network=5.0], id = 889716
-                IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[{INDEX_CASE:_key_PK,_key_PK_proxy}], filters=[<($t1, $t2)], rowType=[RecordType(INTEGER L_ORDERKEY)], projects=[[$t0]], requiredColumns=[{2, 13, 14}], inlineScan=[false], collation=[[2 ASC-nulls-first, 5 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=1.0, io=1.0, network=1.0], id = 889107
+IgniteColocatedSortAggregate(group=[{0}], ORDER_COUNT=[COUNT()], collation=[[0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=12.0, cpu=20.0, memory=23.0, io=2.0, network=14.0], id = {id}
+  IgniteSort(sort0=[$0], dir0=[ASC-nulls-first]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=11.0, cpu=19.0, memory=14.0, io=2.0, network=14.0], id = {id}
+    IgniteProject(O_ORDERPRIORITY=[$1]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.0, cpu=15.0, memory=10.0, io=2.0, network=14.0], id = {id}
+      IgniteCorrelatedNestedLoopJoin(condition=[true], joinType=[inner], variablesSet=[[$cor0]], variablesSet=[[0]], correlationVariables=[[$cor0]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.0, cpu=14.0, memory=10.0, io=2.0, network=14.0], id = {id}
+        IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=9.0], id = {id}
+          IgniteIndexScan(table=[[PUBLIC, ORDERS]], index=[O_OD], filters=[AND(>=($t1, 1993-07-01), <($t1, 1993-10-01))], rowType=[RecordType(INTEGER O_ORDERKEY, VARCHAR O_ORDERPRIORITY)], projects=[[$t0, $t2]], requiredColumns=[{2, 6, 7}], searchBounds=[[null, null, null, null, null, null, RangeBounds [lowerBound=1993-07-01, upperBound=1993-10-01, lowerInclude=true, upperInclude=false], null, null, null, null]], inlineScan=[false], collation=[[6 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
+        IgniteColocatedHashAggregate(group=[{0}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=6.0, cpu=8.0, memory=9.0, io=1.0, network=5.0], id = {id}
+          IgniteProject(i=[true]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=5.0, cpu=7.0, memory=5.0, io=1.0, network=5.0], id = {id}
+            IgniteSortedIndexSpool(readType=[LAZY], writeType=[EAGER], condition=[=($0, $cor0.O_ORDERKEY)], collation=[[0 ASC-nulls-first]], searchBounds=[[ExactBounds [bound=$cor0.O_ORDERKEY]]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=4.0, cpu=6.0, memory=5.0, io=1.0, network=5.0], id = {id}
+              IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=1.0, io=1.0, network=5.0], id = {id}
+                IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[{INDEX_CASE:_key_PK,_key_PK_proxy}], filters=[<($t1, $t2)], rowType=[RecordType(INTEGER L_ORDERKEY)], projects=[[$t0]], requiredColumns=[{2, 13, 14}], inlineScan=[false], collation=[[2 ASC-nulls-first, 5 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=1.0, io=1.0, network=1.0], id = {id}
