+++ modules/calcite/src/test/resources/tpch/q6.plan
@@ -1,3 +1,3 @@
-IgniteColocatedHashAggregate(group=[{}], REVENUE=[SUM($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=3.0, memory=6.0, io=1.0, network=5.0], id = 996526
-  IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=5.0], id = 996525
-    IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_SD], filters=[AND(AND(>=($t3, 1994-01-01), <($t3, 1995-01-01)), AND(>=($t2, 0.05:DECIMAL(4, 2)), <=($t2, 0.07:DECIMAL(4, 2))), <($t0, 24.00))], rowType=[RecordType(DECIMAL(30, 4) $f0)], projects=[[*($t1, $t2)]], requiredColumns=[{6, 7, 8, 12}], searchBounds=[[null, null, null, null, null, null, null, null, null, null, null, null, RangeBounds [lowerBound=1994-01-01, upperBound=1995-01-01, lowerInclude=true, upperInclude=false], null, null, null, null, null]], inlineScan=[false], collation=[[12 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 996515
+IgniteColocatedHashAggregate(group=[{}], REVENUE=[SUM($0)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=3.0, memory=6.0, io=1.0, network=5.0], id = {id}
+  IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=5.0], id = {id}
+    IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_SD], filters=[AND(AND(>=($t3, 1994-01-01), <($t3, 1995-01-01)), AND(>=($t2, 0.05:DECIMAL(4, 2)), <=($t2, 0.07:DECIMAL(4, 2))), <($t0, 24.00))], rowType=[RecordType(DECIMAL(30, 4) $f0)], projects=[[*($t1, $t2)]], requiredColumns=[{6, 7, 8, 12}], searchBounds=[[null, null, null, null, null, null, null, null, null, null, null, null, RangeBounds [lowerBound=1994-01-01, upperBound=1995-01-01, lowerInclude=true, upperInclude=false], null, null, null, null, null]], inlineScan=[false], collation=[[12 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
