+++ modules/calcite/src/test/resources/tpch/variant_q14.plan
@@ -1,8 +1,8 @@
-IgniteProject(PROMO_REVENUE=[/(*(100.00:DECIMAL(5, 2), $0), $1)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.0, cpu=19.0, memory=12.0, io=2.0, network=22.0], id = 2126446
-  IgniteColocatedHashAggregate(group=[{}], agg#0=[SUM($0)], agg#1=[SUM($1)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.0, cpu=18.0, memory=12.0, io=2.0, network=22.0], id = 2126445
-    IgniteProject($f0=[CASE(=(SUBSTRING($1, 1, 5), _UTF-8'PROMO'), *($3, -(1, $4)), 0.0000:DECIMAL(31, 4))], $f1=[*($3, -(1, $4))]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=8.0, cpu=17.0, memory=2.0, io=2.0, network=22.0], id = 2126444
-      IgniteMergeJoin(condition=[=($2, $0)], joinType=[inner], variablesSet=[[]], leftCollation=[[0 ASC-nulls-first]], rightCollation=[[0 ASC-nulls-first]], allowNulls=[{}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=7.0, cpu=16.0, memory=2.0, io=2.0, network=22.0], id = 2126443
-        IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=9.0], id = 2126441
-          IgniteIndexScan(table=[[PUBLIC, PART]], index=[{INDEX_CASE:_key_PK,_key_PK_proxy}], requiredColumns=[{2, 6}], inlineScan=[false], collation=[[2 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = 2122821
-        IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=1.0, io=1.0, network=13.0], id = 2126442
-          IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_PK_SK], filters=[AND(>=($t3, 1995-09-01), <($t3, 1995-10-01))], rowType=[RecordType(INTEGER L_PARTKEY, DECIMAL(15, 2) L_EXTENDEDPRICE, DECIMAL(15, 2) L_DISCOUNT)], projects=[[$t0, $t1, $t2]], requiredColumns=[{3, 7, 8, 12}], inlineScan=[false], collation=[[3 ASC-nulls-first, 4 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=1.0, io=1.0, network=1.0], id = 2122796
+IgniteProject(PROMO_REVENUE=[/(*(100.00:DECIMAL(5, 2), $0), $1)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=10.0, cpu=19.0, memory=12.0, io=2.0, network=22.0], id = {id}
+  IgniteColocatedHashAggregate(group=[{}], agg#0=[SUM($0)], agg#1=[SUM($1)]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=9.0, cpu=18.0, memory=12.0, io=2.0, network=22.0], id = {id}
+    IgniteProject($f0=[CASE(=(SUBSTRING($1, 1, 5), _UTF-8'PROMO'), *($3, -(1, $4)), 0.0000:DECIMAL(31, 4))], $f1=[*($3, -(1, $4))]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=8.0, cpu=17.0, memory=2.0, io=2.0, network=22.0], id = {id}
+      IgniteMergeJoin(condition=[=($2, $0)], joinType=[inner], variablesSet=[[]], leftCollation=[[0 ASC-nulls-first]], rightCollation=[[0 ASC-nulls-first]], allowNulls=[{}]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=7.0, cpu=16.0, memory=2.0, io=2.0, network=22.0], id = {id}
+        IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=2.0, memory=1.0, io=1.0, network=9.0], id = {id}
+          IgniteIndexScan(table=[[PUBLIC, PART]], index=[{INDEX_CASE:_key_PK,_key_PK_proxy}], requiredColumns=[{2, 6}], inlineScan=[false], collation=[[2 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=1.0, cpu=1.0, memory=1.0, io=1.0, network=1.0], id = {id}
+        IgniteExchange(distribution=[single]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=3.0, cpu=6.0, memory=1.0, io=1.0, network=13.0], id = {id}
+          IgniteIndexScan(table=[[PUBLIC, LINEITEM]], index=[L_PK_SK], filters=[AND(>=($t3, 1995-09-01), <($t3, 1995-10-01))], rowType=[RecordType(INTEGER L_PARTKEY, DECIMAL(15, 2) L_EXTENDEDPRICE, DECIMAL(15, 2) L_DISCOUNT)], projects=[[$t0, $t1, $t2]], requiredColumns=[{3, 7, 8, 12}], inlineScan=[false], collation=[[3 ASC-nulls-first, 4 ASC-nulls-first, 2 ASC-nulls-first, 5 ASC-nulls-first, 0 ASC-nulls-first]]): rowcount = 1.0, cumulative cost = IgniteCost [rowCount=2.0, cpu=5.0, memory=1.0, io=1.0, network=1.0], id = {id}
